requirements-dev.txt
-r requirements.txt

django-debug-toolbar==2.1
django-extensions==2.2.6

pytest==5.3.5
pytest-django==3.8.0

black==19.10b0
